Add support for first day and non integer workload
The argument fdom will now allow users to specify when entries should
start only from a specific day of the month, as is often the case with
tutor jobs at the beginning of the semester. Moreover, hrs is now a
float so you the user can specify half hours in case of half contracts.